+++ src/game/default/g_ballistics.c
@@ -861,7 +861,7 @@ static void G_LightningProjectile_Think(g_entity_t *self) {
 	}
 
 	VectorMA(start, g_balance_lightning_length->value, forward, end); // resolve end
-	VectorMA(end, 2.0 * sin(g_level.time / 4.0), up, end);
+	VectorMA(end, 2.0 * sinf(g_level.time / 4.0), up, end);
 	VectorMA(end, 2.0 * Randomc(), right, end);
 
 	tr = gi.Trace(start, end, NULL, NULL, self, MASK_CLIP_PROJECTILE | MASK_LIQUID);
